refactor: move Supabase and TanStack files from utils to lib
